requirements.txt
langchain
langchain_community
langchain_core
langchain_groq
langchain_text_splitters
langchain_chroma
langchain-openai
langchain-pinecone
python-dotenv
faiss-cpu
pypdf
